// src/API.ts
export enum AccountCategoryType {
  BUYER = "BUYER",
  SELLER = "SELLER",
}
// ...
export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}
// ...
export type CreateChatRoomMutationVariables = {
  input: CreateChatRoomInput,
  condition?: ModelChatRoomConditionInput | null,
};

export type CreateChatRoomMutation = {
  createChatRoom?:  {
    __typename: "ChatRoom",
    id: string,
    createdAt: string,
    SType: string,
    newMessages?: number | null,
    name?: string | null,
    imageUri?: string | null,
    lastMessage?:  {
      __typename: "Message",
      id: string,
      text?: string | null,
      createdAt: string,
      SType: string,
      rffID?: string | null,
      rfqID?: string | null,
      rfqType?: string | null,
      rffType?: string | null,
      sellOfferID?: string | null,
      requestID?: string | null,
      requestTitle?: string | null,
      requestQty?: number | null,
      packageType?: string | null,
      serviceType?: serviceType | null,
      requestPrice?: number | null,
      serviceImage?: string | null,
      requestFrom?: string | null,
      requestFromImg?: string | null,
      requestTo?: string | null,
      requestToImg?: string | null,
      status?: MessageStatus | null,
      image?: string | null,
      file?: string | null,
      replyToMessageID?: string | null,
      userID: string,
      chatroomID: string,
      updatedAt: string,
    } | null,
    Messages?:  {
      __typename: "ModelMessageConnection",
      items:  Array< {
        __typename: "Message",
        id: string,
        text?: string | null,
        createdAt: string,
        SType: string,
        rffID?: string | null,
        rfqID?: string | null,
        rfqType?: string | null,
        rffType?: string | null,
        sellOfferID?: string | null,
        requestID?: string | null,
        requestTitle?: string | null,
        requestQty?: number | null,
        packageType?: string | null,
        serviceType?: serviceType | null,
        requestPrice?: number | null,
        serviceImage?: string | null,
        requestFrom?: string | null,
        requestFromImg?: string | null,
        requestTo?: string | null,
        requestToImg?: string | null,
        status?: MessageStatus | null,
        image?: string | null,
        file?: string | null,
        replyToMessageID?: string | null,
        userID: string,
        chatroomID: string,
        updatedAt: string,
      } | null >,
      nextToken?: string | null,
    } | null,
    users?:  {
      __typename: "ModelUserChatRoomConnection",
      items:  Array< {
        __typename: "UserChatRoom",
        id: string,
        chatRoomId: string,
        userId: string,
        chatRoom:  {
          __typename: "ChatRoom",
          id: string,
          createdAt: string,
          SType: string,
          newMessages?: number | null,
          name?: string | null,
          imageUri?: string | null,
          updatedAt: string,
          chatRoomLastMessageId?: string | null,
        },
        user:  {
          __typename: "User",
          id: string,
          name: string,
          email: string,
          phone_number?: string | null,
          rating?: number | null,
          accountType?: string | null,
          lastOnlineAt?: number | null,
          lat?: number | null,
          lng?: number | null,
          ledgerBalance?: number | null,
          address?: string | null,
          city?: string | null,
          state?: string | null,
          zipCode?: string | null,
          lga?: string | null,
          website?: string | null,
          incorporateDate?: string | null,
          rcNumber?: string | null,
          totalOrders?: number | null,
          level?: string | null,
          identification?: string | null,
          identificationNumber?: string | null,
          identityDocs?: Array< string | null > | null,
          keyProduct?: string | null,
          country?: string | null,
          inviteCode?: string | null,
          accountCategory?: AccountCategoryType | null,
          title?: string | null,
          logo?: string | null,
          backgroundImage?: string | null,
          images?: Array< string | null > | null,
          businessType?: string | null,
          certifications?: string | null,
          certsDoc?: Array< string | null > | null,
          mainMarkets?: Array< string | null > | null,
          memberShipType?: string | null,
          sellerLevel?: string | null,
          estRevenue?: number | null,
          totalStaff?: string | null,
          responseTime?: string | null,
          languages?: Array< string | null > | null,
          legalRep?: string | null,
          overview?: string | null,
          activeOrder?: number | null,
          createdAt: string,
          updatedAt: string,
        },
        createdAt: string,
        updatedAt: string,
      } | null >,
      nextToken?: string | null,
    } | null,
    updatedAt: string,
    chatRoomLastMessageId?: string | null,
  } | null,
};
// ...
export type CreateUserChatRoomMutationVariables = {
  input: CreateUserChatRoomInput,
  condition?: ModelUserChatRoomConditionInput | null,
};
// ...
export type ProductByDateQueryVariables = {
  SType: string,
  createdAt?: ModelStringKeyConditionInput | null,
  sortDirection?: ModelSortDirection | null,
  filter?: ModelProductFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ProductByDateQuery = {
  productByDate?:  {
    __typename: "ModelProductConnection",
    items:  Array< {
      __typename: "Product",
      id: string,
      createdAt: string,
      SType?: string | null,
      title?: string | null,
      productImage?: string | null,
      image?: string | null,
      images?: Array< string | null > | null,
      description?: string | null,
      productSpec?: string | null,
      rating?: number | null,
      tags?: Array< string | null > | null,
      productCert?: string | null,
      supplyCapacity?: string | null,
      minOrderQty?: string | null,
      packageType?: string | null,
      quantity?: number | null,
      noOfReviews?: number | null,
      transportMode?: string | null,
      placeOrigin?: string | null,
      dateAvailable?: string | null,
      productDocs?: Array< string | null > | null,
      productCertDocs?: Array< string | null > | null,
      category?: string | null,
      commodityCategory?: string | null,
      userID: string,
      Reviews?:  {
        __typename: "ModelReviewConnection",
        items:  Array< {
          __typename: "Review",
          id: string,
          createdAt: string,
          SType?: string | null,
          name?: string | null,
          rating?: number | null,
          comment?: string | null,
          forUserID?: string | null,
          userID: string,
          productID: string,
          updatedAt: string,
        } | null >,
        nextToken?: string | null,
      } | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};
// ...
export type GetUserQueryVariables = {
  id: string,
};
// ...
export type ListUserChatRoomsQuery = {
  listUserChatRooms?:  {
    __typename: "ModelUserChatRoomConnection",
    items:  Array< {
      __typename: "UserChatRoom",
      id: string,
      chatRoomId: string,
      userId: string,
      chatRoom:  {
        __typename: "ChatRoom",
        id: string,
        createdAt: string,
        SType: string,
        newMessages?: number | null,
        name?: string | null,
        imageUri?: string | null,
        lastMessage?:  {
          __typename: "Message",
          id: string,
          text?: string | null,
          createdAt: string,
          SType: string,
          rffID?: string | null,
          rfqID?: string | null,
          rfqType?: string | null,
          rffType?: string | null,
          sellOfferID?: string | null,
          requestID?: string | null,
          requestTitle?: string | null,
          requestQty?: number | null,
          packageType?: string | null,
          serviceType?: serviceType | null,
          requestPrice?: number | null,
          serviceImage?: string | null,
          requestFrom?: string | null,
          requestFromImg?: string | null,
          requestTo?: string | null,
          requestToImg?: string | null,
          status?: MessageStatus | null,
          image?: string | null,
          file?: string | null,
          replyToMessageID?: string | null,
          userID: string,
          chatroomID: string,
          updatedAt: string,
        } | null,
        Messages?:  {
          __typename: "ModelMessageConnection",
          nextToken?: string | null,
        } | null,
        users?:  {
          __typename: "ModelUserChatRoomConnection",
          nextToken?: string | null,
        } | null,
        updatedAt: string,
        chatRoomLastMessageId?: string | null,
      },
      user:  {
        __typename: "User",
        id: string,
        name: string,
        email: string,
        phone_number?: string | null,
        rating?: number | null,
        accountType?: string | null,
        lastOnlineAt?: number | null,
        lat?: number | null,
        lng?: number | null,
        ledgerBalance?: number | null,
        address?: string | null,
        city?: string | null,
        state?: string | null,
        zipCode?: string | null,
        lga?: string | null,
        website?: string | null,
        incorporateDate?: string | null,
        rcNumber?: string | null,
        totalOrders?: number | null,
        level?: string | null,
        identification?: string | null,
        identificationNumber?: string | null,
        identityDocs?: Array< string | null > | null,
        keyProduct?: string | null,
        country?: string | null,
        inviteCode?: string | null,
        accountCategory?: AccountCategoryType | null,
        title?: string | null,
        logo?: string | null,
        backgroundImage?: string | null,
        images?: Array< string | null > | null,
        businessType?: string | null,
        certifications?: string | null,
        certsDoc?: Array< string | null > | null,
        mainMarkets?: Array< string | null > | null,
        memberShipType?: string | null,
        sellerLevel?: string | null,
        estRevenue?: number | null,
        totalStaff?: string | null,
        responseTime?: string | null,
        languages?: Array< string | null > | null,
        legalRep?: string | null,
        overview?: string | null,
        activeOrder?: number | null,
        Orders?:  {
          __typename: "ModelOrderConnection",
          nextToken?: string | null,
        } | null,
        Products?:  {
          __typename: "ModelProductConnection",
          nextToken?: string | null,
        } | null,
        Reviews?:  {
          __typename: "ModelReviewConnection",
          nextToken?: string | null,
        } | null,
        RFQS?:  {
          __typename: "ModelRFQConnection",
          nextToken?: string | null,
        } | null,
        RFFS?:  {
          __typename: "ModelRFFConnection",
          nextToken?: string | null,
        } | null,
        SellOffers?:  {
          __typename: "ModelSellOfferConnection",
          nextToken?: string | null,
        } | null,
        Wishlists?:  {
          __typename: "ModelWishlistConnection",
          nextToken?: string | null,
        } | null,
        Notifications?:  {
          __typename: "ModelNotificationConnection",
          nextToken?: string | null,
        } | null,
        Messages?:  {
          __typename: "ModelMessageConnection",
          nextToken?: string | null,
        } | null,
        ChatRooms?:  {
          __typename: "ModelUserChatRoomConnection",
          nextToken?: string | null,
        } | null,
        createdAt: string,
        updatedAt: string,
      },
      createdAt: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};